pyproject.toml
[project]
name = "lightme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "dashscope>=1.25.15",
    "docx2txt>=0.9",
    "edge-tts>=7.2.8",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai",
    "markdown>=3.10.2",
    "ntplib>=0.4.0",
    "openai>=2.29.0",
    "pypdf>=6.9.2",
    "pywebview>=6.1",
    "pyyaml>=6.0.3",
    "unstructured>=0.21.5",
]
